[av] [LINUXDAR-2180] Switch to new SUSI.
Remove mlmodel from SDDS output.
Only include INSTALL-SET output for dev builds.

+++ av/build.sh
@@ -409,7 +409,15 @@ function build()
         echo "Separate SDDS component"
         [[ -f $SDDS/SDDS-Import.xml ]] || exitFailure $FAILURE_COPY_SDDS_FAILED "Failed to create SDDS-Import.xml"
         cp -a "$SDDS" output/SDDS-COMPONENT || exitFailure $FAILURE_COPY_SDDS_FAILED "Failed to copy Plugin SDDS component to output"
-        cp -a "$INSTALLSET" output/INSTALL-SET || exitFailure $FAILURE_COPY_SDDS_FAILED "Failed to copy Plugin SDDS component to output"
+
+        ## only copy INSTALLSET if it contains lrdata!
+        if [[ -f "$INSTALLSET/files/plugins/av/chroot/susi/distribution_version/version1/lrdata/filerep.dat" ]]
+        then
+            ## Only copy for dev builds - makes it clearer that Prod build have to go to a warehouse or have supplements injected
+            cp -a "$INSTALLSET" output/INSTALL-SET || exitFailure $FAILURE_COPY_SDDS_FAILED "Failed to copy Plugin SDDS component to output"
+        else
+            echo "Not copying INSTALL-SET as it doesn't contain LR data"
+        fi
     else
         echo "Combined installset and SDDS component"
         [[ -f $INSTALLSET/SDDS-Import.xml ]] || exitFailure $FAILURE_COPY_SDDS_FAILED "Failed to create SDDS-Import.xml"
